fix useSubscribeTocRerender
